+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 21.6.5
 - when you are changing the editor font in the settings you will get now a dialog box that
   you might need to restart the application (for [#2163](https://github.com/pbek/QOwnNotes/issues/2163))
+- an attempt was made to improve the editor margin calculation by searching
+  for the word `mono` in the font family and handling those potentially monospaced
+  fonts differently (for [#2163](https://github.com/pbek/QOwnNotes/issues/2163))
 
 ## 21.6.4
 - fixed a problem when turning on the sub-folders panel after it was turned off